refactor: split InteractiveApp shell into hook and layout
Move derived bridge/layout state into useInteractiveAppShell and render
the default overworld shell from InteractiveAppDefaultShell so the
entry component stays a thin scene router.
